Replace minizip-ng with libzip for ZIP64 archive support
Use libzip instead of minizip-ng: simpler CMake integration (single
find_package vs dual-name fallback), available via apt on Ubuntu 24.04
(no build-from-source needed), and equivalent functionality for STORE
entry random access via zip_fseek/zip_fread/zip_ftell.
