+++ src/tests/class_tests/openms/CMakeLists.txt
@@ -64,7 +64,7 @@ foreach(t ${Boost_dependent_tests})
 endforeach()
 
 target_link_libraries(Base64_test ZLIB::ZLIB)
-target_link_libraries(MinizipNG_test ${OPENMS_MINIZIP_TARGET})
+target_link_libraries(Libzip_test libzip::zip)
 target_link_libraries(LPWrapper_test CoinOR::CoinOR)
 target_link_libraries(SpectraSTSimilarityScore_test Eigen3::Eigen)
 target_link_libraries(BinnedSpectrum_test Eigen3::Eigen)
@@ -105,11 +105,11 @@ if (WITH_PARQUET)
   ${OPENMS_ARROW_TARGET}
   ${OPENMS_PARQUET_TARGET}
   )
-  target_link_libraries(MinizipNG_test
+  target_link_libraries(Libzip_test
   ${OPENMS_ARROW_TARGET}
   ${OPENMS_PARQUET_TARGET}
   )
-  target_compile_definitions(MinizipNG_test PRIVATE WITH_PARQUET)
+  target_compile_definitions(Libzip_test PRIVATE WITH_PARQUET)
 endif()
 
 #------------------------------------------------------------------------------
